V2.2.5 Add support for synchronising the processing of asynchronous (bridged) messages.
This allows one process to issue a request (to a transmitter) that will result in a response from a transmitter. The request contains a unique tracking ID that will permit the response to be waited for on the second transmitter. A configurable timeout is used to allow longer running request to be left to run asynchronously and handled via separate logic.